Struct pallet_fragments::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn create(
origin: OriginFor<T>,
proto_hash: Hash256,
metadata: FragmentMetadata<T::AssetId>,
permissions: FragmentPerms,
unique: Option<UniqueOptions>,
max_supply: Option<u64>
) -> DispatchResult

Create a Fragment Definition using an existing Proto-Fragment.

Note: Only the Proto-Fragment’s owner is allowed to create a Fragment Definition using the Proto-Fragment

Arguments

origin - Origin of the extrinsic function

proto_hash - Hash of an existing Proto-Fragment

metadata - Metadata of the Fragment Definition

permissions - Set of Actions (encapsulated in a FragmentPerms bitflag enum) that are allowed to be done to any Fragment Instance when it first gets created from the Fragment Definition that is created in this extrnisic function (e.g edit, transfer etc.)

Note: These allowed set of actions of a created Fragment Instance may change when the Fragment Instance is given to another account ID (see the give extrinsic).

unique (optional) - Whether the Fragment Definiton is unique

max_supply (optional) - Maximum amount of Fragment instances (where each Fragment instance has a different Edition ID) that can be created using the Fragment Definition

pub fn publish(
origin: OriginFor<T>,
fragment_hash: Hash128,
price: u128,
quantity: Option<u64>,
expires: Option<T::BlockNumber>,
amount: Option<u64>
) -> DispatchResult

Put the Fragment Definition fragment_hash on sale. When a Fragment Definition is put on sale, users can create Fragment Instances from it for a fee.

Note: Only the Fragment’s Proto-Fragment’s owner is allowed to put the Fragment on sale

Arguments

origin - Origin of the extrinsic function

fragment_hash - ID of the Fragment Definition to put on sale

price - Price to buy a single Fragment Instance that is created from the *Fragment Definition

quantity (optional) - Maximum amount of Fragment Instances that can be bought

expires (optional) - Block number that the sale ends at (optional)

amount (optional) - If the Fragment instance represents a stack of stackable items (for e.g gold coins or arrows - https://runescape.fandom.com/wiki/Stackable_items), the number of items to top up in the stack of stackable items

pub fn unpublish(origin: OriginFor<T>, fragment_hash: Hash128) -> DispatchResult

Take the Fragment Definition fragment_hash off sale. When a Fragment Definition is put on sale, users can create Fragment Instances from it for a fee.

Note: Only the Fragment’s Proto-Fragment’s owner is allowed to take the Fragment off sale

Arguments

origin - Origin of the extrinsic function

fragment_hash - ID of the Fragment Definition to take off sale

pub fn mint(
origin: OriginFor<T>,
fragment_hash: Hash128,
options: FragmentBuyOptions,
amount: Option<u64>
) -> DispatchResult

Create Fragment instance(s) from the Fragment Definition fragment_hash and assign their ownership to origin

Note: Each created Fragment instance will have a different Edition ID and a Copy ID of “1”.

Note: Only the Fragment Definition’s Proto-Fragment’s owner is allowed to create instance(s) of the Fragment in this extrinsic function.

Arguments

origin - Origin of the extrinsic function

fragment_hash - *ID of the Fragment Definition

options - Enum indicating whether to create one Fragment Instance with custom data attached to it or whether to create multiple Fragment Instances (with no custom data attached)

amount (optional) - If the Fragment Instance(s) represent a stack of stackable items (for e.g gold coins or arrows - https://runescape.fandom.com/wiki/Stackable_items), amount is the number of items to top up in the stack of stackable items

pub fn buy(
origin: OriginFor<T>,
fragment_hash: Hash128,
options: FragmentBuyOptions
) -> DispatchResult

Allows the Caller Account ID origin to create Fragment instance(s) of the Fragment Definition fragment_hash, for a fee. The ownership of the created Fragment instance(s) is assigned to the Caller Account ID.

Note: Each created Fragment instance will have a different Edition ID and a Copy ID of “1”.

Note: The total fee that the buyer (i.e the Caller Account ID origin) must pay is the specified price-per-instance multiplied by the total number of instance(s) that the buyer wants to create. (@karan) This amount will be transferred to the Fragment Definition’s Vault’s Account ID.

Arguments

origin - Origin of the extrinsic function

fragment_hash - ID of the Fragment Definition

options - Enum indicating whether to create one Fragment Instance with custom data attached to it or whether to create multiple Fragment Instances (with no custom data attached)

pub fn give(
origin: OriginFor<T>,
class: Hash128,
edition: u64,
copy: u64,
to: <T::Lookup as StaticLookup>::Source,
new_permissions: Option<FragmentPerms>,
expiration: Option<T::BlockNumber>
) -> DispatchResult

Give the Fragment Instance whose Fragment Definition ID is class, whose Edition ID is edition and whose Copy ID is copy to to.

If the current permitted actions of the Fragment Instance allows for it to be duplicated (i.e if it has the permission FragmentPerms::COPY), then it is duplicated and the duplicate’s ownership is assigned to to. Otherwise, its ownernship is transferred from origin to to.

Note: Only the Fragment Instance’s owner is allowed to give the Fragment Instance

Arguments

origin - Origin of the extrinsic function

class - Fragment Definition ID of the Fragment Instance to give

edition - Edition ID of the Fragment Insance to give

copy - Copy ID of the Fragment instance to give

to - Account ID to give the Fragment instance to

new_permissions (optional) - The permitted set of actions (encapsulated in a FragmentPerms bitflag enum) that the account that is given the Fragment instance can do with it.

Note: new_permissions must be a subset of the current permissions field of the Fragment Instance; therefore, the new_permissions can only be more restrictive (than the current permissions field of the Fragment Instance), never more permissive

expiration (optional) - Block number that the duplicated Fragment Instance expires at. If the Fragment Instance was not duplicated, this parameter is irrelevant.

pub fn create_account(
origin: OriginFor<T>,
class: Hash128,
edition: u64,
copy: u64
) -> DispatchResult

Create the Account ID of the Fragment Instance whose Fragment Definition ID is class, whose Edition ID is edition and whose Copy ID is copy**

Arguments

origin - Origin of the extrinsic function

class - Fragment Definition ID of the Fragment Instance

edition - Edition ID of the Fragment Instance

copy - Copy ID of the Fragment Instance

impl<T: Config> Pallet<T>

pub fn get_vault_id(class_hash: Hash128) -> T::AccountId

Get the Account ID of the Fragment Definition class_hash**

This Account ID is determinstically computed using the Fragment Definition class_hash

pub fn get_fragment_account_id(
class_hash: Hash128,
edition: u64,
copy: u64
) -> T::AccountId

Get the Account ID of the Fragment Instance whose Fragment Definition ID is class_hash, whose Edition ID is edition and whose Copy ID is copy**

This Account ID is determinstically computed using the Fragment Definition ID class_hash, the Edition ID edition and the Copy ID copy

pub fn mint_fragments(
to: &T::AccountId,
fragment_hash: &Hash128,
sale: Option<&PublishingData<T::BlockNumber>>,
options: &FragmentBuyOptions,
quantity: u64,
current_block_number: T::BlockNumber,
expiring_at: Option<T::BlockNumber>,
amount: Option<Compact<u64>>
) -> DispatchResult

Create quantity number of Fragment Instances from the Fragment Definition fragment_hash and assigns their ownership to to

Arguments

to - Account ID to assign ownernship of the created Fragment instances to

fragment_hash - ID of the Fragment Definition

sale - Struct representing a sale of the Fragment Definition, if the Fragment Definition is currently on sale

options - Enum indicating whether to create one Fragment Instance with custom data attached to it or whether to create multiple Fragment Instances (with no custom data attached)

quantity - Number of Fragment Instances to create

current_block_number - Current block number of the Clamor Blockchain

expiring_at (optional) - Block Number that the Fragment Instance will expire at

amount (optional) - If the Fragment Instance(s) represent a stack of stackable items (for e.g gold coins or arrows - https://runescape.fandom.com/wiki/Stackable_items), amount is the number of items to top up in the stack of stackable items