Type Definition pallet_fragments::pallet::CopiesCount

pub type CopiesCount<T: Config> = StorageMap<_GeneratedPrefixForStorageCopiesCount<T>, Identity, (Hash128, u64), Compact<u64>>;

StorageMap that maps a tuple that contains a Fragment Definition ID and an Edition ID to the total number of Fragment Instances that have the Fragment Definition ID and the Edition ID