Type Definition pallet_fragments::pallet::Owners

pub type Owners<T: Config> = StorageDoubleMap<_GeneratedPrefixForStorageOwners<T>, Identity, Hash128, Twox64Concat, T::AccountId, Vec<(Compact<InstanceUnit>, Compact<InstanceUnit>)>>;

StorageDoubleMap that maps a Fragment Definition and a Clamor Account ID to a list of Fragment Instances of the Fragment Definition that is owned by the Clamor Account ID

This storage item stores the exact same thing as Inventory, except that the primary key and the secondary key are swapped

Footnotes:

Notice this pulls from memory (and deserializes (scale)) the whole Vec<_,_>, this is on purpose as it should not be too big.