Crate pallet_fragments

This pallet fragments performs logic related to Fragment Definitions and Fragment Instances

IMPORTANT STUFF TO KNOW:

Fragment Definition

A Fragment Definition is created using a Proto-Fragment (see pallet protos). A Fragment Definition’s ID can be determinstically computed using its Proto-Fragment hash and its metadata struct FragmentMetadata.

A Fragment Definition is essentially a digital asset that can be used to enhance the user experience in a game or application, like an in-game item or user account. A Fragment has its own storage, metadata and digital life on its own.

Fragment Instance

A Fragment Instance is created from a Fragment Definition.

It is a digital asset that can be used to enhance the user experience in a game or application, like an in-game item or user account.

Each Fragment Instance also has an edition number.

Therefore, a Fragment Instance can be uniquely identified using its Fragment Definition’s ID, its Edition ID and its Copy ID.

The Copy ID allows us to distinguish a Fragment Instance that has the same Fragment Definition ID and the same Edition ID.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

FragmentDefinition

Struct of a Fragment Definition

FragmentInstance

Struct of a Fragment Instance

FragmentMetadata

Struct of a Fragment Definition’s Metadata

GetDefinitionsParams

Data Type used to Query and Filter for Fragment Definitions

GetInstancesParams

Data Type used to Query and Filter for Fragment Instances

PublishingData

Struct representing a sale of the Fragment Definition .

UniqueOptions

TODO Enum that represents the settings for a Fragment Definition whose Fragment instance(s) must contain unique data when created

Enums

FragmentBuyOptions

Enum indicating whether to create one Fragment Instance with custom data attached to it or whether to create multiple Fragment Instances (with no custom data attached)

Traits

WeightInfo

Weight functions needed for pallet_fragments.