Module pallet_fragments::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

CopiesCount

StorageMap that maps a tuple that contains a Fragment Definition ID and an Edition ID to the total number of Fragment Instances that have the Fragment Definition ID and the Edition ID

DataHashMap

StorageMap that maps a Fragment Definition ID and a Number to a Data Hash

DataHashMapIndex

StorageMap that maps a Fragment Definition ID to the total number of “Numbers” (see DataHashMap to understand what “Numbers” means) that fall under it

Definition2SecondarySales

StorageNMap that maps a Fragment Instance’s Fragment Definition ID, Edition ID and Copy ID to a SecondarySaleData struct

Definitions

StorageMap that maps a Fragment Definition ID (which is determinstically computed using its Proto-Fragment hash and its metadata struct DefinitionMetadata) to a FragmentDefinition struct

EditionsCount

StorageMap that maps a Fragment Definition ID to the total number of unique Edition IDs found in the Fragment Instances that have the aforementioned Fragment Definition ID

Expirations

StorageMap that maps the Block Number to a list of Fragment Instances that expire on that Block (note: each FI in the list is represented as a tuple that contains the Fragment Instance’s Fragment Definition ID, the Fragment Instance’s Edition ID and the Fragment Instance’s Copy ID)

Fragments

StorageNMap that maps a Fragment Instance’s Fragment Definition ID, Edition ID and Copy ID to a FragmentInstance struct

Inventory

StorageDoubleMap that maps a Fragnova Account ID and a Fragment Definition to a list of Fragment Instances of the Fragment Definition that is owned by the Fragnova Account ID

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Owners

StorageDoubleMap that maps a Fragment Definition and a Fragnova Account ID to a list of Fragment Instances of the Fragment Definition that is owned by the Fragnova Account ID

Proto2Fragments

StorageMap that maps a Proto-Fragment to a list of Fragment Definitions that were created using the aforementioned Proto-Fragment

Publishing

StorageMap that maps a Fragment Definition ID to a PublishingData struct (of the aforementioned Fragment Definition)

UniqueData2Edition

StorageMap that maps a Fragment Definition and a Unique Data Hash to an Existing Edition of the aforementioned Fragment Definition